Move grants structure
