+++ grantly/src/components/GrantList.js
@@ -6,6 +6,7 @@ import { connect } from "react-redux";
 import Grant from "./Grant";
 import Loader from "react-loader-spinner";
 import { fetchApi } from "../actions";
+// ".src/actions/index.js";
 
 // Styles
